% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_correlations.R
\name{print.PartialCorrelations}
\alias{print.PartialCorrelations}
\title{Print method for PartialCorrelations objects}
\usage{
\method{print}{PartialCorrelations}(x, ...)
}
\arguments{
\item{x}{A PartialCorrelations object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for PartialCorrelations objects
}
