% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rbm.R
\name{print.RBM}
\alias{print.RBM}
\title{Print method for RBM objects}
\usage{
\method{print}{RBM}(x, ...)
}
\arguments{
\item{x}{An RBM object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for RBM objects
}
