% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rbm.R
\name{summary.RBM}
\alias{summary.RBM}
\title{Summary method for RBM objects}
\usage{
\method{summary}{RBM}(object, ...)
}
\arguments{
\item{object}{An RBM object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for RBM objects
}
